refactor: 파일구조 변경
- detail-modal과 tree-memo를 `tree-memo-list-node` 폴더 내로 이동
- TreeBaseMemoNode -> TreeMemoBaseNode 로 이름 변경
