+++ apps/client/src/shared/components/memo-list-view/components/tree-view/components/index.ts
@@ -1,3 +1,3 @@
 export { default as CustomEdge } from './custom-edge';
-export { default as TreeBaseMemoNode } from './tree-base-memo-node/tree-base-memo-node';
+export { default as TreeMemoBaseNode } from './tree-memo-base-node/tree-memo-base-node';
 export { default as TreeMemoListNode } from './tree-memo-list-node/tree-memo-list-node';
