+++ apps/client/src/shared/components/memo-list-view/components/tree-view/components/tree-memo-list-node/detail-modal/detail-modal.css.ts
