+++ apps/client/src/shared/components/memo-list-view/components/tree-view/components/tree-memo-list-node/tree-memo/tree-memo.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
 
-import DetailModal from '@shared/components/modals/detail-modal/detail-modal';
 import { LABEL_COLOR_BY_TEXT } from '@shared/constants/label-match';
 import useSingleAndDoubleClick from '@shared/hooks/use-single-and-double-click';
 import { LabelTextType } from '@shared/types/label-type';
 import { StructureMemoTypes } from '@shared/types/memo-info-type';
 
+import DetailModal from '../detail-modal/detail-modal';
 import { useDetailMemo } from './apis/queries';
 
 import * as styles from './tree-memo.css';
@@ -16,7 +16,6 @@ interface TreeMemoProps {
 
 const TreeMemo = ({ memo }: TreeMemoProps) => {
   const { memoId, title, content, labelList } = memo;
-
   const [isModalOpen, setIsModalOpen] = useState(false);
 
   // 모달이 열릴 때만 API 호출
@@ -33,20 +32,16 @@ const TreeMemo = ({ memo }: TreeMemoProps) => {
       sourceMemoTitleList: [],
     },
   } = useDetailMemo({ memoId, enabled: isModalOpen });
-  const labelName = labelList[0]?.name ?? ('라벨없음' as LabelTextType);
+  const labelName = labelList[0]?.name ?? '라벨없음';
   const labelColor = LABEL_COLOR_BY_TEXT[labelName as LabelTextType];
 
   const handleModalOpenChange = (open: boolean) => {
     setIsModalOpen(open);
   };
 
-  const handleModalOpen = () => {
-    setIsModalOpen(true);
-  };
-
   const handleClick = useSingleAndDoubleClick({
-    handleSingleClick: handleModalOpen,
-    handleDoubleClick: handleModalOpen,
+    handleSingleClick: () => handleModalOpenChange(true),
+    handleDoubleClick: () => handleModalOpenChange(true),
   });
 
   return (
